Refactor out latency assertions into a common harness function
This is the start of the intended pattern I'm thinking about for these
tests writ large. These integration tests should be making assertions
about behavior given a certain configuration, and the minutae of
validating said behavior should be in a common location that we can
easily update if and when behavioral changes are made.