Add validation and modeling of the GraphQL responses in the test harness (#6)
This is a pretty gnarly harness function, enough so to justify its own
unit tests. It lets us validate that our responses are query-compliant
and then we can separately parse the body into the loosely-modeled
structs to eventually make assertions about things like the null
ratio and scalar generation logic.

Because we use fixed seeds for generating the GraphQL queries (but let
the responses remain random), we can then hard-code test case assertions
against the response Structs knowing what fields were requested.
This will be mildly brittle if any of the underlying RNG in apollo-smith
or arbitrary ever makes a change that causes them to change the output
we're coding against for a given seed, but I expect those changes are not
going to be made lightly or frequently enough in those crates to require
us to make changes to test cases often.