Refactor test structure and test names
Because of the way unit tests run, having the harness tests within the
module meant that as we added more integration tests the harness unit
tests would run every time. This changes them to be a standalone
integration test that will only run once.

I also forgot that Rust appends "test" on the console output for tests,
so prefixing the functions reads weirdly on the output. I refactored
that while I was moving stuff around.