Fix response generation to populate field aliases
This ends up being a nontrivial change to how response generation works.
In addition to not handling field aliases, our code was not
spec-compliant on how GraphQL queries handle repeated occurrences of a
selection set.

The correct spec behavior is to merge all occurrences of a given
selection before generating the response, such that any occurence of the
object has all the fields requested in aggregate accross all selections.

Consider the following simple query:

```
{
  foo {
    some_alias: bar
  }
  foo {
    bar
  }
}
```

The old code would have returned a response of the form (overwriting the
first generated `foo` struct entirely):

```
{
  "foo": {
    "bar": 1
  }
}
```

The new code properly merges the request:

```
{
  "foo": {
    "bar": 1,
    "some_alias": 1
  }
}
```