Add a test for the default response generation configuration (#14)
This uses a query that has all the possible data types represented to
assert the default configuration behavior of our response generator.

The test schema needed to have some additions made to ensure all the
data types were represented, and our scalar generator for boolean
values was bugged and is now fixed (GraphQL's boolean type is named
"Boolean" not "Bool").