Switch to serde_json_bytes (#28)
This aligns us with the rest of Apollo's Rust tech stack.
Unfortunately, it doesn't fix our high drop cost issue. We're paying a
high cost because we're building up so many intermediate `Value` types
that then all get dropped in place when we use `to_vec` to convert that
into a byte response. If we want to optimize that out (and if it's even
worth our time to try to shave this down further), we'll likely need to
work from a different approach that doesn't build up the whole JSON
object as a DOM before serialization which may be more complexity than
it's worth for this use case.